chore(protocol): move get_type_from_record to protocol
