+++ sn_networking/src/cmd.rs
@@ -25,7 +25,7 @@ use libp2p::{
 use sn_evm::{AttoTokens, PaymentQuote, QuotingMetrics};
 use sn_protocol::{
     messages::{Cmd, Request, Response},
-    storage::{RecordHeader, RecordKind, RecordType},
+    storage::{get_type_from_record, RecordType},
     NetworkAddress, PrettyPrintRecordKey,
 };
 use std::{
@@ -34,7 +34,6 @@ use std::{
     time::Duration,
 };
 use tokio::sync::oneshot;
-use xor_name::XorName;
 
 use crate::target_arch::Instant;
 
@@ -557,33 +556,6 @@ impl SwarmDriver {
         Ok(())
     }
 
-    /// Return the RecordType
-    pub(crate) fn get_type_from_record(record: &Record) -> Result<RecordType> {
-        let key = record.key.clone();
-        let record_key = PrettyPrintRecordKey::from(&key);
-
-        match RecordHeader::from_record(record) {
-            Ok(record_header) => match record_header.kind {
-                RecordKind::Chunk => Ok(RecordType::Chunk),
-                RecordKind::Scratchpad => Ok(RecordType::Scratchpad),
-                RecordKind::Spend | RecordKind::Register => {
-                    let content_hash = XorName::from_content(&record.value);
-                    Ok(RecordType::NonChunk(content_hash))
-                }
-                RecordKind::ChunkWithPayment
-                | RecordKind::RegisterWithPayment
-                | RecordKind::ScratchpadWithPayment => {
-                    error!("Record {record_key:?} with payment shall not be stored locally.");
-                    Err(NetworkError::InCorrectRecordHeader)
-                }
-            },
-            Err(err) => {
-                error!("For record {record_key:?}, failed to parse record_header {err:?}");
-                Err(NetworkError::InCorrectRecordHeader)
-            }
-        }
-    }
-
     pub(crate) fn handle_local_cmd(&mut self, cmd: LocalSwarmCmd) -> Result<(), NetworkError> {
         let start = Instant::now();
         let mut cmd_string;
@@ -667,7 +639,7 @@ impl SwarmDriver {
                 let key = record.key.clone();
                 let record_key = PrettyPrintRecordKey::from(&key);
 
-                let record_type = Self::get_type_from_record(&record)?;
+                let record_type = get_type_from_record(&record)?;
 
                 let result = self
                     .swarm
